+++ src/main/kotlin/me/nobaboy/nobaaddons/features/slayers/CompactSlayerMessages.kt
@@ -85,9 +85,7 @@ object CompactSlayerMessages {
 	}
 
 	private fun send() {
-		if(config.removeLastMessage) {
-			lastMessage?.remove()
-		}
+		if(config.removeLastMessage) lastMessage?.remove()
 		lastMessage = ChatUtils.addAndCaptureMessage(compile(), prefix = false, color = null)
 	}