+++ src/main/kotlin/me/nobaboy/nobaaddons/features/slayers/MiniBossFeatures.kt
@@ -30,7 +30,7 @@ object MiniBossFeatures {
 		if(lastAlert.elapsedSince() < 1.seconds) return
 		if(event.id != EXPLODE || event.volume != 0.6f || event.pitch != 9 / 7f) return
 
-		RenderUtils.drawTitle(tr("nobaaddons.slayers.miniBossAlert.spawned", "MiniBoss Spawned!"), config.miniBossAlert.alertColor, duration = 1.5.seconds, id = "slayer.alert")
+		RenderUtils.drawTitle(tr("nobaaddons.slayers.miniBossAlert.spawned", "MiniBoss Spawned!"), config.miniBossAlert.alertColor, duration = 1.5.seconds, id = "slayer_alert")
 		SoundUtils.dingLowSound.play()
 		lastAlert = Timestamp.now()
 	}