+++ CHANGELOG.md
@@ -14,6 +14,10 @@ is handled internally, and as such things could potentially break.
 - Fishing catch timer HUD element - celeste
 - Option to revert fishing treasure catch messages to the pre-Backwater Bayou format - celeste
 - Option to mute Banshee - nobaboy
+- Compact Sea Creature catch messages - nobaboy
+- Announce Sea Creatures - nobaboy
+  - This is only limited to Legendary and Mythic rarities
+- Hotspot Waypoints - nobaboy
 
 ### Changed
 
@@ -26,6 +30,8 @@ is handled internally, and as such things could potentially break.
 - Various config menu settings not correctly saving - celeste
 - Enchantment tooltip parsing properly works on enchanted books again - celeste
 - Certain config options re-using the same key - celeste
+- Inconsistency with Mineshaft Entrance/Ladder waypoints - nobaboy
+- YACL labels not showing - nobaboy
 
 ## 1.0.0-Beta.1 - 2025-03-18